refactor(docs→drive): move +info shortcut to drive domain
URL resolution is a cross-document capability (docx, sheet, bitable,
wiki, file, etc.) backed by drive.metas.batch_query, so it belongs
under the drive service, not docs.
