+++ shortcuts/drive/shortcuts.go
@@ -24,5 +24,6 @@ func Shortcuts() []common.Shortcut {
 		DriveTaskResult,
 		DriveApplyPermission,
 		DriveSearch,
+		DriveInfo,
 	}
 }
