+++ tests/cli_e2e/drive/drive_info_dryrun_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 Lark Technologies Pte. Ltd.
 // SPDX-License-Identifier: MIT
 
-package docs
+package drive
 
 import (
 	"context"
@@ -13,19 +13,17 @@ import (
 	"github.com/tidwall/gjson"
 )
 
-// TestDocs_InfoDryRun_DocxURL asserts that docs +info --dry-run with a docx URL
+// TestDriveInfoDryRun_DocxURL asserts that drive +info --dry-run with a docx URL
 // plans a single batch_query API call (no wiki unwrap needed).
-func TestDocs_InfoDryRun_DocxURL(t *testing.T) {
-	t.Setenv("LARKSUITE_CLI_APP_ID", "app")
-	t.Setenv("LARKSUITE_CLI_APP_SECRET", "secret")
-	t.Setenv("LARKSUITE_CLI_BRAND", "feishu")
+func TestDriveInfoDryRun_DocxURL(t *testing.T) {
+	setDriveInfoE2EEnv(t)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	t.Cleanup(cancel)
 
 	result, err := clie2e.RunCmd(ctx, clie2e.Request{
 		Args: []string{
-			"docs", "+info",
+			"drive", "+info",
 			"--url", "https://xxx.feishu.cn/docx/doxcnDryRunE2E",
 			"--dry-run",
 		},
@@ -42,19 +40,17 @@ func TestDocs_InfoDryRun_DocxURL(t *testing.T) {
 		"expected batch_query URL, stdout:\n%s", result.Stdout)
 }
 
-// TestDocs_InfoDryRun_WikiURL asserts that docs +info --dry-run with a wiki URL
+// TestDriveInfoDryRun_WikiURL asserts that drive +info --dry-run with a wiki URL
 // plans a two-step flow: get_node then batch_query.
-func TestDocs_InfoDryRun_WikiURL(t *testing.T) {
-	t.Setenv("LARKSUITE_CLI_APP_ID", "app")
-	t.Setenv("LARKSUITE_CLI_APP_SECRET", "secret")
-	t.Setenv("LARKSUITE_CLI_BRAND", "feishu")
+func TestDriveInfoDryRun_WikiURL(t *testing.T) {
+	setDriveInfoE2EEnv(t)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	t.Cleanup(cancel)
 
 	result, err := clie2e.RunCmd(ctx, clie2e.Request{
 		Args: []string{
-			"docs", "+info",
+			"drive", "+info",
 			"--url", "https://xxx.feishu.cn/wiki/wikcnDryRunE2E",
 			"--dry-run",
 		},
@@ -74,19 +70,17 @@ func TestDocs_InfoDryRun_WikiURL(t *testing.T) {
 		"expected batch_query as second step, stdout:\n%s", result.Stdout)
 }
 
-// TestDocs_InfoDryRun_BareTokenWithType asserts that docs +info --dry-run with a
+// TestDriveInfoDryRun_BareTokenWithType asserts that drive +info --dry-run with a
 // bare token and --type plans a single batch_query call.
-func TestDocs_InfoDryRun_BareTokenWithType(t *testing.T) {
-	t.Setenv("LARKSUITE_CLI_APP_ID", "app")
-	t.Setenv("LARKSUITE_CLI_APP_SECRET", "secret")
-	t.Setenv("LARKSUITE_CLI_BRAND", "feishu")
+func TestDriveInfoDryRun_BareTokenWithType(t *testing.T) {
+	setDriveInfoE2EEnv(t)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
 	t.Cleanup(cancel)
 
 	result, err := clie2e.RunCmd(ctx, clie2e.Request{
 		Args: []string{
-			"docs", "+info",
+			"drive", "+info",
 			"--url", "doxcnBareToken",
 			"--type", "docx",
 			"--dry-run",
@@ -101,4 +95,12 @@ func TestDocs_InfoDryRun_BareTokenWithType(t *testing.T) {
 	require.Equal(t, "/open-apis/drive/v1/metas/batch_query",
 		gjson.Get(result.Stdout, "api.0.url").String(),
 		"expected batch_query URL, stdout:\n%s", result.Stdout)
+}
+
+func setDriveInfoE2EEnv(t *testing.T) {
+	t.Helper()
+	t.Setenv("LARKSUITE_CLI_CONFIG_DIR", t.TempDir())
+	t.Setenv("LARKSUITE_CLI_APP_ID", "drive_info_e2e_app")
+	t.Setenv("LARKSUITE_CLI_APP_SECRET", "drive_info_e2e_secret")
+	t.Setenv("LARKSUITE_CLI_BRAND", "feishu")
 }
